Big cleanup.
Cleaned up code in oCharging object, oViewController object, scEIDDraw,
scDefineWeapon, and scPlayerCreate. Was mostly commenting, but found
some bugs along the way. Fixed a bug involving the sprite for the charge
beam not changing for certain beams.

Also included an .exe for testing with Windows 10 in compatibility mode
for Windows 7. Will try to include with each commit.